fix(ai): move AIVisionError/AIVisionFailureKind to domain to fix arch boundary
App layer must not import from infra. Moved the vision error types to
src.domain.exceptions.ai_exceptions; infra/ai_vision_errors.py is now a
re-export shim for backward compatibility.
